Partition the datasets by target host.
We don't have a single place to put 40TB, but split across

- duke.neuro.polymtl.ca
- beluga.computecanada.ca:/projects
- cedar.computecanada.ca:/projects
- graham.computecanada.ca:/projects

we have enough, temporarily. This script round-robin allocates the subjects
to each disk. It's not smart enough to figure out which server it's
running on, you have to tell it, but that's pretty easy.

This also commits the files created by @alexfoias that:
- cache the results of our target query
- list the subjects already downloaded to smb://duke.neuro.polymtl.ca/mri/ukbiobank/

The first file (selected_subjects.csv) was created by a script, I think,
but I don't have it. Can we include it here?